further cleanup of unused variables, parameters, etc.  This commit addresses almost all warnings and errors generated when using the -Wextra comiler flag

+++ drivers/classic/src/close_files.c
@@ -29,7 +29,7 @@ close_files(filep_struct         *filep,
 **********************************************************************/
 {
     extern option_struct options;
-    int                  filenum;
+    size_t               filenum;
 
     /**********************
        Close All Input Files
