+++ vic_run/src/soil_carbon_balance.c
@@ -52,8 +52,8 @@ soil_carbon_balance(soil_con_struct   *soil_con,
 {
     extern option_struct       options;
     extern global_param_struct global_param;
-    int                        i;
-    int                        Nnodes;
+    size_t                     i;
+    size_t                     Nnodes;
     double                    *dZ;
     double                    *dZCum;
     double                     dZTot;
